removeTagFromVideo
